--- Assets/PlayerTracking.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4cac5a4363332a54a887704fd5ff73bf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
